+++ README.md
@@ -1,6 +1,7 @@
 # studiopulse
 StudioPulse is an offline desktop application for media project tracking and file path management. Built using React, Django REST Framework, and Electron, it enables teams to collaborate efficiently, track assets, and manage local files within structured project workflows.
-
+NOTE: Activate your venv by 
+ source ./venv/Scripts/activate in bash
 
 # Add permission
 1) Create superuser for application
