+++ apps/task_management/views.py
@@ -2,7 +2,7 @@ from django.http import HttpResponse
 from rest_framework import viewsets
 from rest_framework import  permissions
 from .models import Project, ProjectComment, Shot, ShotAssociation, Comment
-
+from apps.core.security_manager import IsAuthenticated, PublicReadOnly
 from .serializers import (
     ProjectSerializer,
     ProjectCommentSerializer,
@@ -17,26 +17,26 @@ def index(request):
 
 
 class ProjectViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     queryset = Project.objects.all()
     serializer_class = ProjectSerializer
 
 class ProjectCommentViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     queryset = ProjectComment.objects.all()
     serializer_class = ProjectCommentSerializer
 
 class ShotViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     queryset = Shot.objects.all()
     serializer_class = ShotSerializer
 
 class ShotAssociationViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     queryset = ShotAssociation.objects.all()
     serializer_class = ShotAssociationSerializer
 
 class CommentViewSet(viewsets.ModelViewSet):
-    permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [IsAuthenticated]
     queryset = Comment.objects.all()
     serializer_class = CommentSerializer
